Email: fix the mail bigger than 1M can't downloaded on Exchange account.
The exchange update the database through contentresolve, and it is IPC
it will use bundle to deliver the data. The bundle has the size limit,
so it will throw the TransactionTooLargeException, lead Exchange can't
update database, so the mail can't downloaded.

Compress the mail content before put the data to bundle, and decompress
the data in EmailProvider.

Change-Id: I60462146dd4a3b2772b29d8a1ac3532dbca1d159
CRs-fixed: 943810